feat: init wizard UX, list command, CI/CD, dep upgrades, sync progress
- ASCII art mascot + interactive init wizard with storage/provider setup
- brainjar list command with --json support
- data_dir config (default ~/.brainjar) for centralized DB storage
- Sync progress bars via indicatif (file count + current file)
- CI workflow (Linux/macOS/Windows matrix, clippy + test)
- Release workflow (v* tags, cross-compiled binaries, GitHub Release)
- Dep upgrades: rusqlite 0.39, reqwest 0.13, rustyline 18
- Removed personal references from README
- Added mascot image + Buy Me a Coffee badge
- release-plz config for conventional commits + auto changelog
- Branch protection docs (.github/BRANCH_PROTECTION.md)
- All 185 tests passing, clippy clean

+++ src/fuzzy.rs
@@ -61,7 +61,7 @@ pub fn correct_query(conn: &Connection, query: &str) -> Result<(String, Vec<(Str
     // Load vocabulary into memory (small — ~7K words for 276 docs)
     let mut stmt = conn.prepare("SELECT word, frequency FROM vocabulary")?;
     let vocab: Vec<(String, usize)> = stmt
-        .query_map([], |row| Ok((row.get::<_, String>(0)?, row.get::<_, usize>(1)?)))?
+        .query_map([], |row| Ok((row.get::<_, String>(0)?, row.get::<_, i64>(1)? as usize)))?
         .filter_map(|r| r.ok())
         .collect();
